Feature(#60): 그룹 웹 소켓 - 게임시작 메시지 처리 기능 추가
