+++ group_service/src/main/java/io/urdego/group_service/api/controller/group/websocket/EventType.java
@@ -1,0 +1,7 @@
+package io.urdego.group_service.api.controller.group.websocket;
+
+public enum EventType {
+    PARTICIPANT,
+    READY,
+    START
+}
