+++ group_service/src/main/java/io/urdego/group_service/api/controller/group/websocket/GroupMessageController.java
@@ -1,10 +1,10 @@
 package io.urdego.group_service.api.controller.group.websocket;
 
-import io.urdego.group_service.api.controller.group.websocket.dto.GroupMemberStatusResponse;
-import io.urdego.group_service.api.controller.group.websocket.dto.GroupStatusResponse;
-import io.urdego.group_service.api.controller.group.websocket.dto.MessageRequest;
-import io.urdego.group_service.api.controller.group.websocket.dto.MessageResponse;
-import io.urdego.group_service.api.service.group.GroupService;
+import io.urdego.group_service.api.controller.group.websocket.request.MessageRequest;
+import io.urdego.group_service.api.controller.group.websocket.response.GameStartResponse;
+import io.urdego.group_service.api.controller.group.websocket.response.GroupStatusResponse;
+import io.urdego.group_service.api.controller.group.websocket.response.GroupMemberStatusResponse;
+import io.urdego.group_service.api.controller.group.websocket.response.MessageResponse;
 import io.urdego.group_service.api.service.groupMember.GroupMemberService;
 import io.urdego.group_service.domain.entity.groupMember.GroupMemberRole;
 import lombok.RequiredArgsConstructor;
@@ -15,6 +15,7 @@ import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.stereotype.Controller;
 
 import java.util.List;
+
 
 @Controller
 @RequiredArgsConstructor
@@ -30,23 +31,23 @@ public class GroupMessageController {
 
         return switch (request.eventType()) {
             case PARTICIPANT -> {
-                String nickname = messageDataReader.read(request.data().get("nickname"));
-                GroupMemberRole role = GroupMemberRole.valueOf(messageDataReader.read(request.data().get("role")));
+                String nickname = messageDataReader.readString(request.data().get("nickname"));
+                GroupMemberRole role = GroupMemberRole.valueOf(messageDataReader.readString(request.data().get("role")));
                 groupMemberService.addMember(groupId, nickname, role);
 
                 List<GroupMemberStatusResponse> groupMemberStatusResponseList = groupMemberService.getStatus(groupId);
                 yield new MessageResponse<>(GroupStatusResponse.of(groupMemberStatusResponseList));
             }
             case READY -> {
-                String nickname = messageDataReader.read(request.data().get("nickname"));
+                String nickname = messageDataReader.readString(request.data().get("nickname"));
                 groupMemberService.ready(groupId, nickname);
 
                 List<GroupMemberStatusResponse> groupMemberStatusResponseList = groupMemberService.getStatus(groupId);
                 yield new MessageResponse<>(GroupStatusResponse.of(groupMemberStatusResponseList));
             }
             case START -> {
-                log.info("EVENT TYPE : START");
-                yield new MessageResponse("hello world!");
+                Long gameId = messageDataReader.readLong(request.data().get("gameId"));
+                yield new MessageResponse<>(GameStartResponse.of(gameId));
             }
         };
     }
