--- group_service/src/main/java/io/urdego/group_service/api/controller/group/websocket/dto/EventType.java
@@ -1,7 +1,0 @@
-package io.urdego.group_service.api.controller.group.websocket.dto;
-
-public enum EventType {
-    PARTICIPANT,
-    READY,
-    START
-}
