--- group_service/src/main/java/io/urdego/group_service/api/controller/group/websocket/dto/MessageRequest.java
@@ -1,9 +1,0 @@
-package io.urdego.group_service.api.controller.group.websocket.dto;
-
-import java.util.Map;
-
-public record MessageRequest(
-        EventType eventType,
-        Map<String, Object> data
-        ) {
-}
