--- group_service/src/main/java/io/urdego/group_service/api/controller/group/websocket/dto/MessageResponse.java
@@ -1,6 +1,0 @@
-package io.urdego.group_service.api.controller.group.websocket.dto;
-
-public record MessageResponse<T>(
-        T data
-) {
-}
