+++ group_service/src/main/java/io/urdego/group_service/api/controller/group/websocket/response/GameStartResponse.java
@@ -1,0 +1,10 @@
+package io.urdego.group_service.api.controller.group.websocket.response;
+
+public record GameStartResponse(
+        Long gameId,
+        String action
+) {
+    public static GameStartResponse of(Long gameId) {
+        return new GameStartResponse(gameId, "startGame");
+    }
+}
