+++ group_service/src/main/java/io/urdego/group_service/api/controller/group/websocket/response/MessageResponse.java
@@ -1,0 +1,6 @@
+package io.urdego.group_service.api.controller.group.websocket.response;
+
+public record MessageResponse<T>(
+        T data
+) {
+}
